Avoid module to delete yourself

+++ common/install.sh
@@ -22,9 +22,9 @@ dafm() {
 ui_print " "
 ui_print "* Found conflicting module!"
 ui_print "* Deleting $line..."
-rm -rf $line
+[ ! $line = $MODPATH ] && rm -rf $line
 line=${line//modules/modules_update}
-[ -d $line ] && rm -rf $line; done
+[ -d $line ] && [ ! $line = $MODPATH ] && rm -rf $line; done
 }
 
 # All to ttf
